refactor: constants 위치 변경
