+++ src/main/java/lotto/constants/message/InputError.java
@@ -1,4 +1,4 @@
-package lotto.domain.constants.message;
+package lotto.constants.message;
 
 public enum InputError implements Message {
     DUPLICATE_LOTTO_NUMBER("[ERROR] 로또의 번호 조합에 중복이 숫자가 발견되었습니다.");
