+++ src/main/java/lotto/constants/value/LottoRule.java
@@ -1,4 +1,4 @@
-package lotto.domain.constants.value;
+package lotto.constants.value;
 
 public enum LottoRule {
     COMBINATION_LENGTH(6),
